[Refactoring ⚙️] comment 관련 로직 훅으로 완전 분리
